[PathFind] ver.2 작업 중 (.Bake 등) * "Study.asmdef"를 추가하여 작업. 에디터에서만 작동

+++ Assets/Script/Editor/MapSampling/EditMapTileJob.cs
@@ -159,7 +159,7 @@ namespace Script.Data
         public int  GridKey;
         public int  TileKey;
         public long NaviMask;
-        public int  LinkMask;
+        public ushort  LinkMask;
         public int  RenderMask;
 
         public EditMapTileData(EditMapTileData visit_tile, int add_link_mask)
@@ -167,7 +167,7 @@ namespace Script.Data
             GridKey   = visit_tile.GridKey;
             TileKey   = visit_tile.TileKey;
             NaviMask  = visit_tile.NaviMask;
-            LinkMask  = visit_tile.LinkMask | add_link_mask;
+            LinkMask  = (ushort)(visit_tile.LinkMask | add_link_mask);
             RenderMask = visit_tile.RenderMask;
         }
         public readonly float3 GetTilePivot()
